# Conflicts:
#	src/client/BL/Saleman/SalemanSaleblservice/SelloutBillMakeController.java

+++ src/server/hibernate/BuyinEntity.java
@@ -5,24 +5,24 @@ import javax.persistence.*;
 @Entity
 @Table(name = "BUYIN", schema = "PIS", catalog = "")
 public class BuyinEntity {
-    private Long kind;
+    private Double kind;
     private String keyno;
     private String note;
     private String oper;
-    private Long ischeck;
-    private Long isred;
+    private Double ischeck;
+    private Double isred;
     private String provider;
     private String base;
     private String goodsoutlist;
-    private Long sumall;
+    private Double sumall;
 
     @Basic
     @Column(name = "KIND", nullable = true, length = 20)
-    public Long getKind() {
+    public Double getKind() {
         return kind;
     }
 
-    public void setKind(Long kind) {
+    public void setKind(Double kind) {
         this.kind = kind;
     }
 
@@ -58,21 +58,21 @@ public class BuyinEntity {
 
     @Basic
     @Column(name = "ISCHECK", nullable = true, precision = 0)
-    public Long getIscheck() {
+    public Double getIscheck() {
         return ischeck;
     }
 
-    public void setIscheck(Long ischeck) {
+    public void setIscheck(Double ischeck) {
         this.ischeck = ischeck;
     }
 
     @Basic
     @Column(name = "ISRED", nullable = true, precision = 0)
-    public Long getIsred() {
+    public Double getIsred() {
         return isred;
     }
 
-    public void setIsred(Long isred) {
+    public void setIsred(Double isred) {
         this.isred = isred;
     }
 
@@ -108,11 +108,11 @@ public class BuyinEntity {
 
     @Basic
     @Column(name = "SUMALL", nullable = true, precision = 0)
-    public Long getSumall() {
+    public Double getSumall() {
         return sumall;
     }
 
-    public void setSumall(Long sumall) {
+    public void setSumall(Double sumall) {
         this.sumall = sumall;
     }
