+++ src/server/hibernate/WarningEntity.java
@@ -5,24 +5,24 @@ import javax.persistence.*;
 @Entity
 @Table(name = "WARNING", schema = "PIS", catalog = "")
 public class WarningEntity {
-    private Long kind;
+    private Double kind;
     private String keyno;
     private String note;
     private String oper;
-    private Long ischeck;
-    private Long isred;
+    private Double ischeck;
+    private Double isred;
     private String goodsno;
     private String goodsname;
-    private Long num;
-    private Long warningnum;
+    private Double num;
+    private Double warningnum;
 
     @Basic
     @Column(name = "KIND", nullable = true, precision = 0)
-    public Long getKind() {
+    public Double getKind() {
         return kind;
     }
 
-    public void setKind(Long kind) {
+    public void setKind(Double kind) {
         this.kind = kind;
     }
 
@@ -58,21 +58,21 @@ public class WarningEntity {
 
     @Basic
     @Column(name = "ISCHECK", nullable = true, precision = 0)
-    public Long getIscheck() {
+    public Double getIscheck() {
         return ischeck;
     }
 
-    public void setIscheck(Long ischeck) {
+    public void setIscheck(Double ischeck) {
         this.ischeck = ischeck;
     }
 
     @Basic
     @Column(name = "ISRED", nullable = true, precision = 0)
-    public Long getIsred() {
+    public Double getIsred() {
         return isred;
     }
 
-    public void setIsred(Long isred) {
+    public void setIsred(Double isred) {
         this.isred = isred;
     }
 
@@ -98,21 +98,21 @@ public class WarningEntity {
 
     @Basic
     @Column(name = "NUM", nullable = true, precision = 0)
-    public Long getNum() {
+    public Double getNum() {
         return num;
     }
 
-    public void setNum(Long num) {
+    public void setNum(Double num) {
         this.num = num;
     }
 
     @Basic
     @Column(name = "WARNINGNUM", nullable = true, precision = 0)
-    public Long getWarningnum() {
+    public Double getWarningnum() {
         return warningnum;
     }
 
-    public void setWarningnum(Long warningnum) {
+    public void setWarningnum(Double warningnum) {
         this.warningnum = warningnum;
     }
